norm: tokenize_create_token*.c (separate file)
