+++ src/tokenize_create_token3.c
@@ -13,111 +13,6 @@
 #include <tokenize.h>
 #include <vector.h>
 
-t_token	*create_or_token(size_t *i)
-{
-	t_token	*token;
-
-	token = null_guard(malloc(sizeof(t_token)),
-			PROGRAM_NAME, "create_or_token().");
-	token->category = T_OR;
-	token->content = NULL;
-	(*i)++;
-	return (token);
-}
-
-t_token	*create_and_token(size_t *i)
-{
-	t_token	*token;
-
-	token = null_guard(malloc(sizeof(t_token)),
-			PROGRAM_NAME, "create_and_token().");
-	token->category = T_AND;
-	token->content = NULL;
-	(*i)++;
-	return (token);
-}
-
-t_token	*create_pipe_token(void)
-{
-	t_token	*token;
-
-	token = null_guard(malloc(sizeof(t_token)),
-			PROGRAM_NAME, "create_pipe_token().");
-	token->category = T_PIPE;
-	token->content = NULL;
-	return (token);
-}
-
-t_token	*create_lparen_token(size_t *i)
-{
-	t_token	*token;
-
-	token = null_guard(malloc(sizeof(t_token)),
-			PROGRAM_NAME, "create_lparen_token().");
-	token->category = T_L_PAREN;
-	token->content = NULL;
-	(*i)++;
-	return (token);
-}
-
-t_token	*create_rparen_token(size_t *i)
-{
-	t_token	*token;
-
-	token = null_guard(malloc(sizeof(t_token)),
-			PROGRAM_NAME, "create_rparen_token().");
-	token->category = T_R_PAREN;
-	token->content = NULL;
-	(*i)++;
-	return (token);
-}
-
-t_token	*create_in_redirect_token(void)
-{
-	t_token	*token;
-
-	token = null_guard(malloc(sizeof(t_token)),
-			PROGRAM_NAME, "create_in_redirect_token().");
-	token->category = T_IN_REDIRECT;
-	token->content = NULL;
-	return (token);
-}
-
-t_token	*create_out_redirect_token(void)
-{
-	t_token	*token;
-
-	token = null_guard(malloc(sizeof(t_token)),
-			PROGRAM_NAME, "create_out_redirect_token().");
-	token->category = T_OUT_REDIRECT;
-	token->content = NULL;
-	return (token);
-}
-
-t_token	*create_append_redirect_token(size_t *i)
-{
-	t_token	*token;
-
-	token = null_guard(malloc(sizeof(t_token)),
-			PROGRAM_NAME, "create_append_redirect_token().");
-	token->category = T_APPEND_REDIRECT;
-	token->content = NULL;
-	(*i)++;
-	return (token);
-}
-
-t_token	*create_here_doc_token(size_t *i)
-{
-	t_token	*token;
-
-	token = null_guard(malloc(sizeof(t_token)),
-			PROGRAM_NAME, "create_here_doc_token().");
-	token->category = T_HERE_DOC;
-	token->content = NULL;
-	(*i)++;
-	return (token);
-}
-
 static int	handling_quote(const char *line, size_t *i, t_vector *vector)
 {
 	const char	quote = line[*i];
